mudanca na estrutura dos arquivos, refatoracoes de validacoes estao em helpers/storeValidate, mudanca do app.ts para app.js, remocao das funcionalidades typescript em geral
