+++ package.json
@@ -2,26 +2,23 @@
   "name": "encontrar-lojas-proximidades",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "src/app.js",
   "scripts": {
-    "start": "ts-node src/app.ts",
-    "dev": "ts-node-dev src/app.ts",
+    "start": "node src/app.js",
+    "dev": "nodemon src/app.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Brendo Brito",
   "license": "ISC",
   "dependencies": {
-    "@types/express": "^5.0.0",
     "axios": "^1.7.7",
     "dotenv": "^16.4.5",
     "express": "^4.21.1",
     "mongodb": "^6.9.0",
     "mongoose": "^8.7.2",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.6.3",
     "winston": "^3.15.0"
   },
   "devDependencies": {
-    "ts-node-dev": "^2.0.0"
+    "nodemon": "^3.0.0"
   }
 }
