+++ src/services/viaCepService.js
@@ -5,7 +5,7 @@ const getAddressFromCep = async(cep) => {
         const response = await axios.get(`https://viacep.com.br/ws/${cep}/json/`);
         return response.data;
     }catch(erro){
-        throw new Error('Erro ao consultar o ViaCEP.');
+        throw new Error('Erro ao buscar o CEP.');
     }
 };
