--- tsconfig.json
@@ -1,16 +1,0 @@
-{
-    "compilerOptions": {
-      "target": "ES6",                         
-      "module": "commonjs",                    
-      "strict": true,                          
-      "esModuleInterop": true,                 
-      "skipLibCheck": true,                    
-      "forceConsistentCasingInFileNames": true,
-      "outDir": "./dist",                      
-      "rootDir": "./src",                      
-      "resolveJsonModule": true,               
-      "moduleResolution": "node"               
-    },
-    "include": ["src/**/*"],
-    "exclude": ["node_modules"]
-  }
